refactor: deleted check publishers & consumers

+++ stock/events/location/LocationDeletedConsumer.ts
@@ -16,20 +16,27 @@ export class LocationDeletedConsumer extends CoreConsumer<LocationConsumerReq> {
           const data = JSON.parse(msg.content.toString());
           console.log(`Received message from ${this.exchange} using routing key: ${this.routingKey}`);
 
-          const deletedItem = await locationController.datamapper.delete(data.id);
+          const checkIfUsed = await locationController.datamapper.checkIfNotNull("ean", data.id);
 
           if (!process.env.REDIS_HOST) {
             throw new Error("Redis host must be set")
           }
 
-          console.log("Location blockage deleted successfully");
-
           const redis = RedisManager.getCmdInstance(process.env.REDIS_HOST, 6379);
           await redis.connect();
 
-          if (deletedItem) {
-            await redis.addResponse({ eventID: data.eventID, success: true });
-          } else if (!deletedItem ) {
+          if (checkIfUsed.length === 0) {
+            const deletedItem = await locationController.datamapper.delete(data.id);
+
+            if (deletedItem) {
+              console.log("Location type deleted successfully");
+              await redis.addResponse({ eventID: data.eventID, success: true });
+            } else {
+              console.log("Location type deletion failed.");
+              await redis.addResponse({ eventID: data.eventID, success: false });
+            }
+          } else {
+            console.log("Location type deletion failed.");
             await redis.addResponse({ eventID: data.eventID, success: false });
           }
