+++ stock/events/location_blockage/LocationBlockageDeletedConsumer.ts
@@ -1,7 +1,7 @@
 import { CoreConsumer, RedisManager, RoutingKeys } from "@zencorp/engrenages";
 import { LocationBlockageConsumerReq } from "../interfaces/location/LocationBlockageConsumerReq";
 import { Channel, ConsumeMessage } from "amqplib";
-import { locationBlockageController } from "../../app/controllers/index.controllers";
+import { locationBlockageController, locationController } from "../../app/controllers/index.controllers";
 
 export class LocationBlockageDeletedConsumer extends CoreConsumer<LocationBlockageConsumerReq> {
   readonly routingKey = RoutingKeys.LocationBlockageDeleted;
@@ -20,20 +20,27 @@ export class LocationBlockageDeletedConsumer extends CoreConsumer<LocationBlocka
           const data = JSON.parse(msg.content.toString());
           console.log(`Received message from ${this.exchange} using routing key: ${this.routingKey}`);
 
-          const deletedItem = await locationBlockageController.datamapper.delete(data.id);
+          const checkIfUsed = await locationController.datamapper.checkIfUsed("location_blockage_name", data.name);
 
           if (!process.env.REDIS_HOST) {
             throw new Error("Redis host must be set")
           }
 
-          console.log("Location blockage deleted successfully");
-
           const redis = RedisManager.getCmdInstance(process.env.REDIS_HOST, 6379);
           await redis.connect();
 
-          if (deletedItem) {
-            await redis.addResponse({ eventID: data.eventID, success: true });
-          } else if (!deletedItem ) {
+          if (checkIfUsed.length === 0) {
+            const deletedItem = await locationBlockageController.datamapper.delete(data.id);
+
+            if (deletedItem) {
+              console.log("Location blockage deleted successfully");
+              await redis.addResponse({ eventID: data.eventID, success: true });
+            } else {
+              console.log("Location blockage deletion failed.");
+              await redis.addResponse({ eventID: data.eventID, success: false });
+            }
+          } else {
+            console.log("Location blockage deletion failed.");
             await redis.addResponse({ eventID: data.eventID, success: false });
           }
