+++ stock/events/product/ProductDeletedConsumer.ts
@@ -1,7 +1,7 @@
 import { CoreConsumer, RedisManager, RoutingKeys } from "@zencorp/engrenages";
-import { ProductConsumerReq } from "../interfaces/product/ProductConsumerConsumerReq";
+import { ProductConsumerReq } from "../interfaces/product/ProductConsumerReq";
 import { Channel, ConsumeMessage } from "amqplib";
-import { productController } from "../../app/controllers/index.controllers";
+import { locationController, productController } from "../../app/controllers/index.controllers";
 
 export class ProductDeletedConsumer extends CoreConsumer<ProductConsumerReq> {
   readonly routingKey = RoutingKeys.ProductDeleted;
@@ -20,20 +20,27 @@ export class ProductDeletedConsumer extends CoreConsumer<ProductConsumerReq> {
           const data = JSON.parse(msg.content.toString());
           console.log(`Received message from ${this.exchange} using routing key: ${this.routingKey}`);
 
-          const deletedItem = await productController.datamapper.delete(data.id);
+          const checkIfUsed = await locationController.datamapper.checkIfNotNull("ean", data.id);
 
           if (!process.env.REDIS_HOST) {
             throw new Error("Redis host must be set")
           }
 
-          console.log("Product deleted successfully");
-
           const redis = RedisManager.getCmdInstance(process.env.REDIS_HOST, 6379);
           await redis.connect();
 
-          if (deletedItem) {
-            await redis.addResponse({ eventID: data.eventID, success: true });
-          } else if (!deletedItem ) {
+          if (checkIfUsed.length === 0) {
+            const deletedItem = await productController.datamapper.delete(data.id);
+
+            if (deletedItem) {
+              console.log("Product deleted successfully");
+              await redis.addResponse({ eventID: data.eventID, success: true });
+            } else {
+              console.log("Product deletion failed.");
+              await redis.addResponse({ eventID: data.eventID, success: false });
+            }
+          } else {
+            console.log("Product deletion failed.");
             await redis.addResponse({ eventID: data.eventID, success: false });
           }
