+++ stock/events/product_blockage/ProductBlockageDeletedConsumer.ts
@@ -1,6 +1,6 @@
 import { CoreConsumer, RedisManager, RoutingKeys } from "@zencorp/engrenages";
 import { ProductBlockageConsumerReq } from "../interfaces/product/ProductBlockageConsumerReq";
-import { productBlockageController } from "../../app/controllers/index.controllers";
+import { productBlockageController, productController } from "../../app/controllers/index.controllers";
 import { Channel, ConsumeMessage } from "amqplib";
 
 export class ProductBlockageDeletedConsumer extends CoreConsumer<ProductBlockageConsumerReq> {
@@ -20,20 +20,27 @@ export class ProductBlockageDeletedConsumer extends CoreConsumer<ProductBlockage
           const data = JSON.parse(msg.content.toString());
           console.log(`Received message from ${this.exchange} using routing key: ${this.routingKey}`);
 
-          const deletedItem = await productBlockageController.datamapper.delete(data.id);
+          const checkIfUsed = await productController.datamapper.checkIfUsed("product_blockage_name", data.name);
 
           if (!process.env.REDIS_HOST) {
             throw new Error("Redis host must be set")
           }
 
-          console.log("Product blockage type deleted successfully");
-
           const redis = RedisManager.getCmdInstance(process.env.REDIS_HOST, 6379);
           await redis.connect();
 
-          if (deletedItem) {
-            await redis.addResponse({ eventID: data.eventID, success: true });
-          } else if (!deletedItem ) {
+          if (checkIfUsed.length === 0) {
+            const deletedItem = await productBlockageController.datamapper.delete(data.id);
+
+            if (deletedItem) {
+              console.log("Product blockage deleted successfully");
+              await redis.addResponse({ eventID: data.eventID, success: true });
+            } else {
+              console.log("Product blockage deletion failed.");
+              await redis.addResponse({ eventID: data.eventID, success: false });
+            }
+          } else {
+            console.log("Product blockage deletion failed.");
             await redis.addResponse({ eventID: data.eventID, success: false });
           }
